domainから実装を分離、成功時のJSONを定義
