+++ app/internal/infrastructure/router.go
@@ -1,7 +1,7 @@
 package infrastructure
 
 import (
-	"github.com/TTT0420/golangCleanArch/internal/domain/repository"
+	"github.com/TTT0420/golangCleanArch/internal/infrastructure/repository"
 	"github.com/TTT0420/golangCleanArch/internal/interface/handler"
 	"github.com/TTT0420/golangCleanArch/internal/usecase"
 	"github.com/gin-gonic/gin"
